+++ src/core/drive/progress_bar.ts
@@ -10,13 +10,16 @@ export class ProgressBar {
         display: block;
         top: 0;
         left: 0;
+        width: 100%;
         height: 3px;
         background: #0076ff;
         z-index: 2147483647;
+        transform-origin: 0 0;
+        transform: scaleX(0) translateZ(0);
         transition:
-          width ${ProgressBar.animationDuration}ms ease-out,
+          transform ${ProgressBar.animationDuration}ms ease-out,
           opacity ${ProgressBar.animationDuration / 2}ms ${ProgressBar.animationDuration / 2}ms ease-in;
-        transform: translate3d(0, 0, 0);
+        will-change: transform, opacity;
       }
     `
   }
@@ -68,7 +71,7 @@ export class ProgressBar {
   }
 
   installProgressElement() {
-    this.progressElement.style.width = "0"
+    this.progressElement.style.transform = "scaleX(0) translateZ(0)"
     this.progressElement.style.opacity = "1"
     document.documentElement.insertBefore(this.progressElement, document.body)
     this.refresh()
@@ -102,7 +105,8 @@ export class ProgressBar {
 
   refresh() {
     requestAnimationFrame(() => {
-      this.progressElement.style.width = `${10 + this.value * 90}%`
+      const scale = (10 + this.value * 90) / 100
+      this.progressElement.style.transform = `scaleX(${scale}) translateZ(0)`
     })
   }