+++ src/tests/functional/form_submission_tests.ts
@@ -38,6 +38,11 @@ test("test standard form submission renders a progress bar", async ({ page }) =>
 
   await waitUntilSelector(page, ".turbo-progress-bar")
   assert.ok(await hasSelector(page, ".turbo-progress-bar"), "displays progress bar")
+  assert.match(
+    await page.locator(".turbo-progress-bar").evaluate((element) => (element as HTMLDivElement).style.transform),
+    /^scaleX\(.+\) translateZ\(0(px)?\)$/,
+    "animates progress with transform"
+  )
 
   await nextBody(page)
   await waitUntilNoSelector(page, ".turbo-progress-bar")